Adds files to plot _data files created by onelayergui
Scripts to plot _data files from onelayergui. The sample files give an
illustration of the function in action and the different parameters that
can be changed. qcmplotter is basically a gui version of the
plotmultiplelines script. There are currently some issues with loading
files, but you can get an idea of what it does.